Fixed voiding of wrong patient HIV staging information.
For every patient mastercard loaded, staging information is cached in
the VueX store, however this cached data is never invalidated and only updated
when a patient with staging information is loaded. In cases where a patient
without staging information is loaded that other patient staging information was
still in the cache. So an update of staging information was resulting in the
existing information being voided first before saving of the new information.